Updated routines for better stability
Version 0.3.7 seemed to stop polling silently, guessing because of the way timers were triggered. This updated use a 'timer.Every <interval>'  to always trigger a poll with exception if it is in back off state.
Updated the deprecated webrequest method and also provided default timeout of 10 seconds.
The backoff routine is there to stop polling for a while if steam stops accepting requests.
I've stepped the version but you may want to review/change/refactor if you decide to use these changes.